Make glyphSet param consistently optional for pens
The following pens have glyphSet as a required param to their
constructor:

  - ControlBoundsPen, BoundsPen
  - CocoaPen
  - PointInsidePen
  - QtPen
  - ReportLabPen
  - SVGPathPen
  - T2CharStringPen
  - TTGlyphPen
  - WxPen

All other pens make it optional (defaulting to None):

  - AreaPen
  - FilterPen, ContourFilterPen, ReverseContourPen
  - MomentsPen
  - PerimeterPen
  - PointToSegmentPen, SegmentToPointPen, GuessSmoothPointPen,
    ReverseContourPointPen
  - RecordingPen, DecomposingRecordingPen
  - StatisticsPen
  - TeePen
  - TransformPen

Of the above pens that require glyphSet (to be passed into the parent
constructor for DecomposingPen--a superclass of BasePen), only
TTGlyphPen actually uses self.glyphSet and addComponent (a method added
by DecomposingPen). Several pens override addComponent or call it on
pens they own (such as TeePen, which calls it on all of its subpens).

The fact that so many pens (which on the surface seem like they may need
to support addComponent--like AreaPen, StatisticsPen, etc.) make
glyphSet optional indicates that it should be safe to make it more
uniformly optional. This way, it isn't so cumbersome to construct an
SVGPathPen, for example.

Note that PointInsidePen cannot have glyphSet made optional, because it
is followed by a required param:

    def __init__(self, glyphSet, testPoint, evenOdd=False):

Reordering the params to accommodate would be a breaking change.