otTables_test: get XML dumps for expected MarkBasePos split subtables
This tests will fail because of this bug: https://github.com/googlefonts/noto-source/issues/145

The mark records' Class values of the second subtable should be updated to match the new class count
(the two split subtable contain half of the original mark classes). Otherwise the base records would
implicitly reference the wrong or non-existent mark classes...
The fix is in the following commit.