[glyf] if comp uses anchors compute firstPt-secondPt offset after applying transform
Fixes https://github.com/fonttools/fonttools/issues/1556

When a component uses firstPt/secondPt reference anchor points instead of XY offsets,
and the component also has a transform, fonttools is incorrectly computing its bounding box.
This is because we are computing the translation offset between firstPt and secondPt before
applying the 2x2 scale/rotation/shear transform. By the time we do the translation, the
offset is now incorrect.

We need to compute the translation offset after we have applied the 2x2 transform.