setup.py: optionally build cython extension
If Cython is already pre-installed, we opportunistically try to compile the
cu2qu extension module, and if it fails (e.g. compiler is not available), then
we silently fall back to pure-python mode.

If user explicitly requests --with-cython (or sets 'FONTTOOLS_WITH_CYTHON=1' env
variable), then we add cython to setup_requires, and raise an error if
compilation fails.