Distinguish between axis tag and id.
Inside an OpenType font an axis is referred to by its index. In ttx
the axis tag was used as an id but the axis tag is not required to be
unique. Explicitly create and use an axis id which is separate from the
axis tag. This allows for proper handling of fonts with multiple axes
which specify the same tag.

Note that this explicitly does not touch the non-fvar axis tags in the
STAT table, which are a separate set of tags.