[varLib] Fix instantiating some SinglePos subtables
Under certain conditions, instantiating GPOS SinglePos subtables results
in all zero values. This seems to be caused by the merger reassigning
the output Value to a new ValueRecord before merging it, but the
“things” it is being merged with is a list with a single item which is
the output Value itself, so it gets overridden to zero before merging.

The above is probably gibberish since I don’t really understand much of
varLib, but the fix seems to copy the old Value to when creating a new
one before merging.

Without this fix the added test will have its SinglePos values
incorrectly instantiated to zero.