Update svgLib test for now non-negative arc radius values
Missed this test in previous commit.

Test is testing that parsing extraction of individual parameters
works as specified when spaces are omitted. Such as signaled when
a new number is specified with a leading sign +/-

Test previously used
     "M1-2A3-4-1.0 01.5.7",
where the first arc radius value was '3' and the second was '-4'.

Now that we are forcing radius values to be non-negative using abs()
the value returned in the test is not
     ("arcTo", (3.0, -4.0, -1.0, False, True, (0.5, 0.7))),
but rather
     ("arcTo", (3.0, 4.0, -1.0, False, True, (0.5, 0.7))),

Changed to expect the positive value, nicely the test continues to
test omitting spaces, but now also tests that negative radius values
are forced non-negative.