[feaLib] More efficient inline multisub compilation
Specifically we were always creating a new ChainContextualRule for each
glyph in an input class, even when their replacements were identical,
which means we would later miss the chance to use a format3 encoding,
which would have been smaller.

The implementation here feels... fine, I guess? In fea-rs we don't
decompose the input set, but we also don't support the syntax,

`sub a [b c]' d by e f` which is supported by fontmake (but has not
yet appeared in any font we've handled)