Prune unused regions from HVAR and VVAR when using a direct mapping
This saves a handful of bytes for the common case where a font is
variable but has no variation in its advances (e.g. for some monospace
fonts across `wght`), and we include an empty metrics variation table so
that the shaper can immediately ascertain this without interpolating
phantom points.