[voltToFea] Fix sorting group definitions
In VOLT group definitions can be ordered in any order, but in FEA a
group must be defined before being referenced in another group. We had
some code to sort the groups to fix this, but it was broken in several
ways. We now use the more appropriate TopologicalSorter(), and nested
enums when getting what groups a group references.