[post] Change the separator for duplicate names from "#" to "."
Using "#" in glyph names can be problematic, e.g. it is not valid if
feature files ("#" starts a comment). This commit uses the less
problematic "." instead.

Because dot is a common glyph name separator, if the font has ["A", "A",
"A.1"] names, we will rename the second glyph to A.1, and then the third
glyph will be seen as duplicate and gets renamed to A.1.1, which is
probably undesired. The code have been slightly modified to make sure
the new glyph name does not exist in the font as whole, not only the
already seen glyphs.