[feaLib] Add promoted single substitutions to aalt feature
If a single substitution was promoted to a multiple or ligature
substitutions, we would no longer add it to aalt feature (when
requested) since only single and alternate substations are added there.

This can be considered a regression from lookup promotion. Since a
single-looking substitution wouldn’t occurs in ligature or multiple
substitution unless we promoted it, it should be safe to add these to
allt feature.

Fixes https://github.com/fonttools/fonttools/issues/3845